frontend/requirements.txt
streamlit
opencv-python
pillow
requests
numpy